# docs/reference/api.yaml
swagger: "2.0"
info:
  title: Tsuru
  description: Open source, extensible and Docker-based Platform as a Service (PaaS)
  version: "1.6"
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /1.8/apps/{app}/routable:
    parameters:
      - name: app
        in: path
        required: true
        type: string
        minLength: 1
        description: App name.
    post:
      operationId: AppSetRoutable
      description: Sets a version as routable.
      tags:
        - app
      security:
        - Bearer: []
      consumes:
        - application/json
      parameters:
        - name: setRoutableData
          in: body
          required: true
          schema:
            $ref: "#/definitions/SetRoutableArgs"
      produces:
        - application/json
      responses:
        "200":
          description: App updated
        "400":
          description: Invalid arguments
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorMessage"
        "404":
          description: App not found
          schema:
            $ref: "#/definitions/ErrorMessage"
...
  /1.8/routers/{name}:
    parameters:
      - name: name
        in: path
        required: true
        type: string
        minLength: 1
        description: Router template name.
    put:
      operationId: RouterUpdate
      description: Updates a router template
      consumes:
        - application/json
      parameters:
        - name: router
          required: true
          in: body
          schema:
            $ref: "#/definitions/RouterTemplate"
      responses:
        "200":
          description: Router template updated
          schema:
            $ref: "#/definitions/TeamToken"
        "401":
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorMessage"
        "404":
          description: Router template not found
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
        - router
      security:
        - Bearer: []
    delete:
      operationId: RouterDelete
      description: Deletes a router template
      responses:
        "200":
          description: Router template deleted
          schema:
            $ref: "#/definitions/TeamToken"
        "401":
          description: Unauthorized.
          schema:
            $ref: "#/definitions/ErrorMessage"
        "404":
          description: Router template not found
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
        - router
      security:
        - Bearer: []
definitions:
  RouterTemplate:
    description: Router template
    type: object
    properties:
      name:
        type: string
      type:
        type: string
      config:
        type: object
...
  ErrorMessage:
    description: Error message.
    type: string
...
  TeamToken:
    description: An authorization token associated to a team.
    type: object
    properties:
      token:
        type: string
      token_id:
        type: string
      description:
        type: string
      created_at:
        type: string
        format: date-time
      expires_at:
        type: string
        format: date-time
      last_access:
        type: string
        format: date-time
      creator_email:
        type: string
      team:
        type: string
      roles:
        type: array
        items:
          type: object
          $ref: "#/definitions/RoleInstance"
  RoleInstance:
    description: Association between a role and a context value.
    type: object
    properties:
      name:
        type: string
      contextvalue:
        type: string
...
  SetRoutableArgs:
    type: object
    properties:
      version:
        type: string
      isRoutable:
        type: boolean
...